frontend/requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# LangChain
langchain==0.1.0
langchain-google-genai==1.0.0
langchain-community==0.0.13
langchain-core==0.1.0

# Vector Store
faiss-cpu==1.7.4

# Document Processing
unstructured==0.11.0
python-magic==0.4.27
# For URL loading
requests==2.31.0

# Web Frameworks
streamlit==1.29.0
fastapi==0.109.0
uvicorn[standard]==0.25.0

# API Models
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1